+++ hismatova.elizaveta/S2/PostfixValue.hpp
@@ -6,15 +6,15 @@ namespace hismatova
   class PostfixValue
   {
   private:
-    int value;
+    long long value;
     char op;
     bool isNumber;
   public:
     PostfixValue();
-    explicit PostfixValue(int val);
+    explicit PostfixValue(long long val);
     explicit PostfixValue(char oper);
     bool isNumeric() const;
-    int getValue() const;
+    long long getValue() const;
     char getOperator() const;
     PostfixValue operator+(const PostfixValue& rhs) const;
     PostfixValue operator-(const PostfixValue& rhs) const;
